# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LSP UGM API Documentation",
        "description": "API Documentation for LSP UGM Application",
        "contact": {
            "name": "Administrator",
            "email": "admin@lsp-ugm.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "LSP UGM API Server"
        }
    ],
    "paths": {
...
        "/asesor/biodata/{id}": {
            "get": {
                "tags": [
                    "Asesor"
                ],
                "summary": "Mendapatkan data biodata asesor",
                "description": "Show biodata asesor",
                "operationId": "13832adc2884086705f00d7c73c681ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID asesor",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data asesor ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data Asesor ditemukan"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data asesor tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data Asesor tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Asesor"
                ],
                "summary": "Update data asesor",
                "description": "Update biodata asesor",
                "operationId": "02e858c28c38a7f12ef6c50cb19a3c28",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID asesor",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "_method",
                                    "nama_asesor",
                                    "no_sertifikat",
                                    "no_hp",
                                    "alamat",
                                    "no_ktp",
                                    "jenis_kelamin",
                                    "kebangsaan",
                                    "no_met",
                                    "kode_pos"
                                ],
                                "properties": {
                                    "_method": {
                                        "description": "Method spoofing untuk Laravel",
                                        "type": "string",
                                        "default": "PUT"
                                    },
                                    "nama_asesor": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "no_sertifikat": {
                                        "type": "string",
                                        "example": "CERT123456"
                                    },
                                    "no_hp": {
                                        "type": "string",
                                        "example": "081234567890"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Contoh No. 123"
                                    },
                                    "no_ktp": {
                                        "type": "string",
                                        "example": "3301012345678901"
                                    },
                                    "jenis_kelamin": {
                                        "type": "string",
                                        "example": "Laki-laki"
                                    },
                                    "kebangsaan": {
                                        "type": "string",
                                        "example": "Indonesia"
                                    },
                                    "no_met": {
                                        "type": "string",
                                        "example": "MET12345"
                                    },
                                    "kode_pos": {
                                        "type": "string",
                                        "example": "12345"
                                    },
                                    "kode_registrasi": {
                                        "type": "string",
                                        "example": "REG-2025-001"
                                    },
                                    "tanda_tangan": {
                                        "description": "File tanda tangan",
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "foto_asesor": {
                                        "description": "Foto asesor",
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data asesor berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data Asesor berhasil diupdate"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data asesor tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data Asesor tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validasi gagal"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/asesor/dashboard/{id}": {
            "get": {
                "tags": [
                    "Asesor"
                ],
                "summary": "Mendapatkan data dashboard asesor",
                "description": "Get data asesor for dashboard page",
                "operationId": "180e9d3bb5c7ec05ade257531f39d3df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID asesor",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data asesor ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data Asesor ditemukan"
                                        },
                                        "data": {
                                            "properties": {
                                                "nama_asesor": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email_asesor": {
                                                    "type": "string",
                                                    "example": "john@example.com"
                                                },
                                                "jumlah_asesi": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "jumlah_event": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "jumlah_skema": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "jumlah_kompetensi_teknis": {
                                                    "type": "integer",
                                                    "example": 4
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data asesor tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data Asesor tidak ditemukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "API_KEY",
                "in": "header"
            },
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "API_KEY",
                "in": "header"
            }
        }
    },
...
}